+++ Workflow.md
@@ -80,7 +80,7 @@ npm run ts -- <Путь к модулю с ts-кодом>
 npm run mockServer
 ```
 
-Запускает фейковый сервер для получения с него заранее прописанных в файле `mockData.json` данных.
+Запускает сервер из которого можно получить данные `mockData.json` данных.
 
 #### Запустить redoc
 
@@ -88,15 +88,7 @@ npm run mockServer
 npm run doc
 ```
 
-Запускает на http://localhost:8080 спецификацию в формате OpenAPI 3 (@redocly/cli не использую из-за проблем с запуском на node 16.18.0).
-
-#### Запустить CLI
-
-```bash
-npm run cli
-```
-
-Запускает CLI (по умолчанию вызывается команда `help`).
+Запускает [по ссылке](http://localhost:8080) спецификацию в формате OpenAPI 3 (использую redoc-cli вместо @redocly/cli из-за проблем с запуском на node 16.18.0).
 
 ### Переменные окружения
 
@@ -118,3 +110,64 @@ npm run cli
 ### Запросы к серверу
 
 Файл с запросами `queries.http` (можно отправлять из редактора если есть расширение или из Postman)
+
+### Работа с консольной программой
+
+`npm run ts ./src/cli.ts -- --import './mocks/test.tsv'` <— пример работы **import**
+
+```plaintext
+Подключение к БД MongoDB.
+Соединение было успешно установлено.
+Создан новый фильм: [object Object]
+Создан новый фильм: Криминальное чтиво
+Создан новый фильм: [object Object]
+Создан новый фильм: Выживший
+Создан новый фильм: [object Object]
+Создан новый фильм: Выживший
+Создан новый фильм: [object Object]
+Создан новый фильм: Отель Гранд Будапешт
+Создан новый фильм: [object Object]
+Создан новый фильм: Королевство Полной Луны
+Создан новый фильм: [object Object]
+Создан новый фильм: Королевство Полной Луны
+Создан новый фильм: [object Object]
+Создан новый фильм: Выживший
+Создан новый фильм: [object Object]
+Создан новый фильм: Криминальное чтиво
+Создан новый фильм: [object Object]
+Создан новый фильм: Отель Гранд Будапешт
+Создан новый фильм: [object Object]
+Создан новый фильм: Отель Гранд Будапешт
+10 строк записано в файл !
+Соединение с базой данных было прервано.
+```
+
+`npm run ts ./src/cli.ts -- --help` <– пример работы команды **help**
+
+```plaintext
+Программа для подготовки данных для REST API сервера.
+
+Пример:
+
+    cli.js --<command> [--arguments]
+    
+Команды:
+
+    --version                   # выводит номер версии
+    --help                      # печатает этот текст
+    --import <path>             # импортирует данные из TSV
+```
+
+`npm run ts ./src/cli.ts -- --version` <– пример работы команды **version**
+
+```plaintext
+.env файл был успешно прочитан.
+1.0.1
+```
+
+`npm run ts ./src/cli.ts -- --generate 10 './mocks/demo.tsv' http://localhost:8888/api` <– пример работы **generate**
+
+```plaintext
+.env файл был успешно прочитан.
+Файл с мок данным ./mocks/demo.tsv был успешно создан
+```
