+++ mocks/demo.tsv
@@ -1,0 +1,10 @@
+Выживший	Двое бандитов Винсент Вега и Джулс Винфилд проводят время в философских беседах в перерыве между разборками и «решением проблем» с должниками своего криминального босса Марселласа Уоллеса. Параллельно разворачивается три истории.	2023-01-13T11:24:02.274Z	thriller	2017	4.1	https://what-to-watch.ru/preview/aviator.jpg	https://what-to-watch.ru/movie/aviator.jpg	Тим Рот;Сэмюэл Л. Джексон;Тильда Суинтон;Лукас Хеджес;Фрэнсис Макдорманд;Том Харди;Донал Глисон;Кейт Бланшетт	Квентин Тарантино	144	Кисик Сашулик	atkishin@yandex.ru	relevant.jpg	vnd778csSSa09cdSScD56.jpeg	orange
+Королевство Полной Луны	Подростки Сэм и Сьюзи сбегают ото всех в свой идеальный мир. Атмосферная сказка Уэса Андерсона о первой любви	2023-01-10T11:24:02.275Z	horror	2005	2.1	https://what-to-watch.ru/preview/pulp-fiction.jpg	https://what-to-watch.ru/movie/moonrise-kingdom.jpg	Ума Турман;Джон Траволта;Тим Рот;Сэмюэл Л. Джексон;Тильда Суинтон;Лукас Хеджес;Фрэнсис Макдорманд;Том Харди;Донал Глисон	Уэс Андерсон	102	Кисик Сашулик	disney@gmail.com	aviator.jpg	vnd778csSSa09cdSScD56.jpeg	pink
+Отель Гранд Будапешт	Картина расскажет о легендарном консьерже популярного европейского отеля и проследит за его жизнью между двумя мировыми войнами, а также за его дружбой с юной сотрудницей, которая становится его протеже.	2023-01-11T11:24:02.276Z	horror	2015	3.9	https://what-to-watch.ru/preview/the-grand-budapest-hotel.jpg	https://what-to-watch.ru/movie/the-grand-budapest-hotel.jpg	Оуэн Уилсон;Сирша Ронан;Эдриен Броуди;Брюс Уиллис;Ума Турман;Джон Траволта;Тим Рот;Сэмюэл Л. Джексон;Тильда Суинтон;Лукас Хеджес;Фрэнсис Макдорманд	Квентин Тарантино	80	Кисик Евик	atkvs@skbkontur.ru	relevant.jpg	OG9SaAd2MsLsln8cdBWAi.jpeg	pink
+Отель Гранд Будапешт	Картина расскажет о легендарном консьерже популярного европейского отеля и проследит за его жизнью между двумя мировыми войнами, а также за его дружбой с юной сотрудницей, которая становится его протеже.	2023-01-10T11:24:02.276Z	comedy	1999	1.3	https://what-to-watch.ru/preview/aviator.jpg	https://what-to-watch.ru/movie/relevant.jpg	Кейт Бланшетт	Уэс Андерсон	179	Кисик Темушка	atkishin@yandex.ru	the-grand-budapest-hotel.jpg	8Ckdnk121nCdsa.jpeg	red
+Отель Гранд Будапешт	Охотник Хью Гласс серьезно ранен на неизведанных просторах американского Дикого Запада. Товарищ Хью по отряду покорителей новых земель Джон Фицжеральд предательски оставляет его умирать в одиночестве. Теперь у Гласса осталось только одно оружие — его сила воли. Он готов бросить вызов первобытной природе, суровой зиме и враждебным племенам индейцев, только чтобы выжить и отомстить Фицжеральду.	2023-01-13T11:24:02.276Z	family	2004	4.1	https://what-to-watch.ru/preview/pulp-fiction.jpg	https://what-to-watch.ru/movie/relevant.jpg	Джон Траволта;Тим Рот;Сэмюэл Л. Джексон;Тильда Суинтон;Лукас Хеджес;Фрэнсис Макдорманд;Том Харди;Донал Глисон;Кейт Бланшетт	Мартин Скорсезе	119	Кисик Сашулик	atkvs@skbkontur.ru	the-grand-budapest-hotel.jpg	Ac8sdkw7c9scsCDscssd.jpeg	red
+Отель Гранд Будапешт	Картина расскажет о легендарном консьерже популярного европейского отеля и проследит за его жизнью между двумя мировыми войнами, а также за его дружбой с юной сотрудницей, которая становится его протеже.	2023-01-08T11:24:02.276Z	family	2013	1.8	https://what-to-watch.ru/preview/the-grand-budapest-hotel.jpg	https://what-to-watch.ru/movie/relevant.jpg	Эдриен Броуди;Брюс Уиллис;Ума Турман;Джон Траволта;Тим Рот;Сэмюэл Л. Джексон;Тильда Суинтон;Лукас Хеджес;Фрэнсис Макдорманд;Том Харди;Донал Глисон;Кейт Бланшетт	Квентин Тарантино	147	Кисик Сашулик	vladatkishkin@gmail.com	aviator.jpg	Chd89SD00cdsYrtPysA7.jpeg	cyan
+Выживший	Подростки Сэм и Сьюзи сбегают ото всех в свой идеальный мир. Атмосферная сказка Уэса Андерсона о первой любви	2023-01-13T11:24:02.276Z	romance	2002	3	https://what-to-watch.ru/preview/aviator.jpg	https://what-to-watch.ru/movie/pulp-fiction.jpg	Том Харди;Донал Глисон;Кейт Бланшетт	Мартин Скорсезе	84	Кисик Владик	disney@gmail.com	relevant.jpg	vnd778csSSa09cdSScD56.jpeg	pink
+Авиатор	Получив от отца небольшую фабрику, Говард Хьюз превратил ее в гигантское, фантастически прибыльное предприятие. Став владельцем огромной кинокомпании, он снял самый дорогой для своего времени фильм и покорил сердца прелестнейших голливудских актрис.	2023-01-11T11:24:02.276Z	comedy	2016	1.3	https://what-to-watch.ru/preview/moonrise-kingdom.jpg	https://what-to-watch.ru/movie/aviator.jpg	Леонардо Ди Каприо;Билл Мюрей;Рэйф Файнс;Оуэн Уилсон;Сирша Ронан;Эдриен Броуди;Брюс Уиллис;Ума Турман;Джон Траволта;Тим Рот	Квентин Тарантино	80	Кисик Евик	test@skbkontur.ru	pulp-fiction.jpg	Chd89SD00cdsYrtPysA7.jpeg	blue
+Королевство Полной Луны	Охотник Хью Гласс серьезно ранен на неизведанных просторах американского Дикого Запада. Товарищ Хью по отряду покорителей новых земель Джон Фицжеральд предательски оставляет его умирать в одиночестве. Теперь у Гласса осталось только одно оружие — его сила воли. Он готов бросить вызов первобытной природе, суровой зиме и враждебным племенам индейцев, только чтобы выжить и отомстить Фицжеральду.	2023-01-09T11:24:02.276Z	documentary	2018	4.6	https://what-to-watch.ru/preview/aviator.jpg	https://what-to-watch.ru/movie/aviator.jpg	Том Харди;Донал Глисон;Кейт Бланшетт	Алехандро Гонсалес Иньярриту	175	Кисик Сашулик	test@skbkontur.ru	the-grand-budapest-hotel.jpg	Chd89SD00cdsYrtPysA7.jpeg	cyan
+Криминальное чтиво	Получив от отца небольшую фабрику, Говард Хьюз превратил ее в гигантское, фантастически прибыльное предприятие. Став владельцем огромной кинокомпании, он снял самый дорогой для своего времени фильм и покорил сердца прелестнейших голливудских актрис.	2023-01-13T11:24:02.276Z	horror	2006	1.5	https://what-to-watch.ru/preview/aviator.jpg	https://what-to-watch.ru/movie/aviator.jpg	Леонардо Ди Каприо;Билл Мюрей;Рэйф Файнс;Оуэн Уилсон;Сирша Ронан;Эдриен Броуди;Брюс Уиллис;Ума Турман;Джон Траволта;Тим Рот;Сэмюэл Л. Джексон	Квентин Тарантино	165	Кисик Никита	atkvs@skbkontur.ru	the-grand-budapest-hotel.jpg	8Ckdnk121nCdsa.jpeg	red
