+++ queries.http
@@ -1,7 +1,7 @@
 ###
 
 # Фильмы
-## Добавить новый фильм
+## Добавить фильм
 POST http://localhost:8000/movies/create HTTP/1.1
 Accept: application/json
 Content-Type: application/json
@@ -31,7 +31,7 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRla0B2ay5jb20iLCJpZ
 
 ###
 
-## Редактирование карточки фильма
+## Редактирование фильма
 PATCH http://localhost:8000/movies/63c29a0483e4f0e1a55550c7 HTTP/1.1
 Accept: application/json
 Content-Type: application/json
@@ -60,13 +60,13 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRla0B2ay5jb20iLCJpZ
 
 ###
 
-## Удаление карточки фильма
+## Удаление фильма
 DELETE http://localhost:8000/movies/63c295e5b577c08e2073c25f HTTP/1.1
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRpa0B2ay5jb20iLCJpZCI6IjYzYzE5NjJiNzRiNjJiYWZmMDdkNGFjZSIsImlhdCI6MTY3MzYzMTMzMCwiZXhwIjoxNjczODA0MTMwfQ.7GpFXSTgboQaaDDFIZUWEpbRqxm-KwPd6jasPvQ4M-U
 
 ###
 
-## Получить список фильмов
+## Получить всех фильмов (можно ограничить выдачу через query limit)
 GET http://localhost:8000/movies HTTP/1.1
 Accept: application/json
 
@@ -78,13 +78,13 @@ Accept: application/json
 
 ###
 
-## Получение детальной информации по фильму
+## Получение информации по фильму
 GET http://localhost:8000/movies/63c29870298bc7e1ffc50056 HTTP/1.1
 Accept: application/json
 
 ###
 
-## Получение промо-фильма
+## Получение информации по промо фильму
 GET http://localhost:8000/movies/promo HTTP/1.1
 Accept: application/json
 
@@ -121,13 +121,13 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRla0B2ay5jb20iLCJpZ
 ###
 
 # Комментарии
-## Получить список комментариев для определённого фильма
+## Получить список комментариев фильма
 GET http://localhost:8000/movies/63c1969174b62baff07d4adb/comments HTTP/1.1
 Accept: application/json
 
 ###
 
-## Добавить новый комментарий к фильму
+## Добавить комментарий к фильму
 POST http://localhost:8000/comments HTTP/1.1
 Content-Type: application/json
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRpa0B2ay5jb20iLCJpZCI6IjYzYzE5NjJiNzRiNjJiYWZmMDdkNGFjZSIsImlhdCI6MTY3MzYzMTMzMCwiZXhwIjoxNjczODA0MTMwfQ.7GpFXSTgboQaaDDFIZUWEpbRqxm-KwPd6jasPvQ4M-U
@@ -141,7 +141,7 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRpa0B2ay5jb20iLCJpZ
 ###
 
 # Пользователи
-## Создание нового пользователя (без аватарки => будет дефолтной)
+## Создание пользователя (без аватарки => будет дефолтной)
 POST http://localhost:8000/users/register HTTP/1.1
 Content-Type: multipart/form-data; boundary=DataToLogin
 
@@ -221,7 +221,7 @@ Content-Type: image/jpeg
 
 ###
 
-## Вход в закрытую часть приложения
+## Вход в закрытую часть приложения (получаем токен с которым делаем остальные запросы)
 POST http://localhost:8000/users/login HTTP/1.1
 Content-Type: application/json
 
@@ -235,3 +235,58 @@ Content-Type: application/json
 ## Проверка токена пользователя
 GET http://localhost:8000/users/login HTTP/1.1
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRla0B2ay5jb20iLCJpZCI6IjYzYzE5OTdkNzRiNjJiYWZmMDdkNGIyYSIsImlhdCI6MTY3MzY5NjgwNSwiZXhwIjoxNjczODY5NjA1fQ.iUYM9pzCIjMZtGGoSJmLgdgO_Y-uwpc_IceoVvvBcI4
+
+###
+
+# Всякие невалидные запросы
+## Создаем фильм без какого-то обязательного поля
+POST http://localhost:8000/movies/create HTTP/1.1
+Accept: application/json
+Content-Type: application/json
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRla0B2ay5jb20iLCJpZCI6IjYzYzE5OTdkNzRiNjJiYWZmMDdkNGIyYSIsImlhdCI6MTY3MzYzMjE0NSwiZXhwIjoxNjczODA0OTQ1fQ.QOSlczvhMOLunlV2pwF6ZGap8sJ4nK3bEtN6p1ZOPHQ
+
+{
+  "description": "Получив от отца небольшую фабрику, Говард Хьюз превратил ее в гигантское, фантастически прибыльное предприятие. Став владельцем огромной кинокомпании, он снял самый дорогой для своего времени фильм и покорил сердца прелестнейших голливудских актрис.",
+  "publishingDate": "2005-05-02T23:59:33.903Z",
+  "genre": "drama",
+  "releaseYear": 2005,
+  "rating": 7,
+  "commentsCount": 290,
+  "previewPath": "https://what-to-watch.ru/preview/aviator.jpg",
+  "moviePath": "https://what-to-watch.ru/movie/aviator.jpg",
+  "actors": [
+    "Леонардо Ди Каприо",
+    "Кейт Бланшетт",
+    "Джуд Лоу"
+  ],
+  "director": "Мартин Скорсезе",
+  "duration": 169,
+  "posterPath": "aviator.jpg",
+  "backgroundImagePath": "OG9SaAd2MsLsln8cdBWAi.jpeg",
+  "backgroundColor": "pink"
+}
+
+###
+## Актеры идут не списком
+POST http://localhost:8000/movies/create HTTP/1.1
+Accept: application/json
+Content-Type: application/json
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRla0B2ay5jb20iLCJpZCI6IjYzYzE5OTdkNzRiNjJiYWZmMDdkNGIyYSIsImlhdCI6MTY3MzYzMjE0NSwiZXhwIjoxNjczODA0OTQ1fQ.QOSlczvhMOLunlV2pwF6ZGap8sJ4nK3bEtN6p1ZOPHQ
+
+{
+  "title": "Авиатор",
+  "description": "Получив от отца небольшую фабрику, Говард Хьюз превратил ее в гигантское, фантастически прибыльное предприятие. Став владельцем огромной кинокомпании, он снял самый дорогой для своего времени фильм и покорил сердца прелестнейших голливудских актрис.",
+  "publishingDate": "2005-05-02T23:59:33.903Z",
+  "genre": "drama",
+  "releaseYear": 2005,
+  "rating": 7,
+  "commentsCount": 290,
+  "previewPath": "https://what-to-watch.ru/preview/aviator.jpg",
+  "moviePath": "https://what-to-watch.ru/movie/aviator.jpg",
+  "actors": "Кисик",
+  "director": "Мартин Скорсезе",
+  "duration": 169,
+  "posterPath": "aviator.jpg",
+  "backgroundImagePath": "OG9SaAd2MsLsln8cdBWAi.jpeg",
+  "backgroundColor": "pink"
+}
